read settings from config ifle
