+++ 5-AccessControl/1-call-api-roles/API/TodoListAPI/Infrastructure/AuthorizationPolicies.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace TodoListAPI.Infrastructure
+{
+    /// <summary>
+    /// Wrapper class the contain all the authorization policies available in this application.
+    /// </summary>
+    public static class AuthorizationPolicies
+    {
+        public const string AssignmentToTaskUserRoleRequired = "AssignmentToTaskUserRoleRequired";
+        public const string AssignmentToTaskAdminRoleRequired = "AssignmentToTaskAdminRoleRequired";
+    }
+}
