+++ 5-AccessControl/1-call-api-roles/API/TodoListAPI/appsettings.json
@@ -4,6 +4,10 @@
     "Domain": "Enter the domain of your Azure AD tenant, e.g. 'contoso.onmicrosoft.com'",
     "TenantId": "Enter the tenant ID",
     "ClientId": "Enter the Client ID (aka 'Application ID')",
+    "Scopes": {
+      "Read": ["TodoList.Read", "TodoList.ReadWrite"],
+      "Write": ["TodoList.ReadWrite"]
+    },
     "Roles": {
       "TaskAdmin": "TaskAdmin",
       "TaskUser": "TaskUser"
@@ -14,7 +18,7 @@
     "LogLevel": {
       "Default": "Information",
       "Microsoft": "Warning",
-      "Microsoft.Hosting.Lifetime": "Information"
+      "Microsoft.Identity": "Information"
     }
   },
   "AllowedHosts": "*"
