+++ 5-AccessControl/1-call-api-roles/SPA/src/app/app.component.html
@@ -4,6 +4,7 @@
     <div class="toolbar-spacer"></div>
 
     <a mat-button [routerLink]="['todo-view']">TodoList</a>
+    <a mat-button [routerLink]="['dashboard']">Dashboard</a>
 
     <button mat-raised-button *ngIf="!loginDisplay" (click)="login()">Login</button>
     <button mat-raised-button color="accent" *ngIf="loginDisplay" (click)="logout()">Logout</button>
